==> 03_pt_quality.ipynb <==
import pandas as pd
import numpy as np

#open list of stops
stops = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/06_Data Delfi/Delfi/stops_epsg3035.csv')
#stops

#add row for number of departures
stops['departures'] = 0
#stops

stop_times = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/06_Data Delfi/Delfi/stop_times.txt', header=0)
stop_times

stop_id_counts = stop_times['stop_id'].value_counts()

for stop_id, count in stop_id_counts.items():
    stops.loc[stops['stop_id'] == stop_id, 'departures'] += count
    
stops    

#remove stops with 0 departures
stops = stops[stops['departures']>0]
stops

stops.to_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/06_Data Delfi/Delfi/stops_departurecount.csv')

#Now: processing 1km grid cells to allow for overlaying with stops in QGIS later
cells_temp = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/11_Mode Choice/1km_Raumtyp_Slope.csv', header=0, usecols = ['cellname','ycoord','xcoord'])


#coordinates currently denominate south-western corner. To treat cells as points, shift coordinates 500m north and east
cells_temp['ycoord']=cells_temp['ycoord']+500
cells_temp['xcoord']=cells_temp['xcoord']+500

cells_temp

#save as csv for qgis
cells_temp.to_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/06_Data Delfi/Delfi/1kmcells_mp.csv')

# ~~~ number of stops is transferred from stops to 1km gridcells (2km radius around center) in QGIS ~~~ #

#add number of stops to csv
cells = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/11_Mode Choice/1km_Raumtyp_Slope.csv')
stops = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/11_Mode Choice/1km_ptquality.csv')


cells['departures'] = cells['cellname'].map(stops.set_index('cellname')['departures_sum'])
cells

cells.to_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/11_Mode Choice/1km_Raumtyp_Slope_ptqual.csv')
