==> 21_NL_car.ipynb <==
import pandas as pd
import numpy as np
import biogeme.database as db
import biogeme.biogeme as bio
from biogeme.expressions import Beta, Variable
from biogeme.nests import OneNestForNestedLogit, NestsForNestedLogit
from biogeme import models

pd.set_option('display.max_columns', None)
pd.set_option('display.min_rows', 20)
pd.set_option('display.max_rows', 100)

# IMPORT SAMPLES (created in 10_data_processing)
#only import one of the sets of samples below
#imports both databases for model estimation as well as for later simulation/validation

#missing from import: 'GITTER_SO_500m','GITTER_SO_1km','PRAEZISION_SO','GITTER_ZO_500m','GITTER_ZO_1km','PRAEZISION_ZO','vpedrad','Raumtyp_SO_12','Raumtyp_SO_21','Raumtyp_SO_22','Raumtyp_ZO_12','Raumtyp_ZO_21','Raumtyp_ZO_22','slope_SO','slope_ZO','slope_SOZO'
#also not importing: one dummy variable level each (age 4 included for interaction with slope)

# SAMPLE: Complete
d_biogeme_complete = pd.read_csv('D:/Processed Data/ModeChoice_input.csv', usecols = ['choice','W_GEW','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_complete = db.Database('ModeChoice-MiD_complete', d_biogeme_complete)
#d_biogeme_complete

# SAMPLE: 100k
d_biogeme_100000 = pd.read_csv('F:/Processed Data/ModeChoice_input_100000.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_100000 = db.Database('ModeChoice-MiD_100000', d_biogeme_100000)
d_biogeme_100000

# SAMPLE: k-fold cross validation (k=5)

#k=1
d_biogeme_k1 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_sim_k1.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k1 = db.Database('ModeChoice-MiD_k1', d_biogeme_k1)

d_biogeme_k2345 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_input_k2345.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k2345 = db.Database('ModeChoice-MiD_k2345', d_biogeme_k2345)

#k=2
d_biogeme_k2 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_sim_k2.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k2 = db.Database('ModeChoice-MiD_k2', d_biogeme_k2)

d_biogeme_k1345 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_input_k1345.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k1345 = db.Database('ModeChoice-MiD_k1345', d_biogeme_k1345)

#k=3
d_biogeme_k3 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_sim_k3.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k3 = db.Database('ModeChoice-MiD_k3', d_biogeme_k3)

d_biogeme_k1245 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_input_k1245.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k1245 = db.Database('ModeChoice-MiD_k1245', d_biogeme_k1245)

#k=4
d_biogeme_k4 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_sim_k4.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k4 = db.Database('ModeChoice-MiD_k4', d_biogeme_k4)

d_biogeme_k1235 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_input_k1235.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k1235 = db.Database('ModeChoice-MiD_k1235', d_biogeme_k1235)

#k=5
d_biogeme_k5 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_sim_k5.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k5 = db.Database('ModeChoice-MiD_k5', d_biogeme_k5)

d_biogeme_k1234 = pd.read_csv('F:/Processed Data/kfcv/ModeChoice_input_k1234.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_k1234 = db.Database('ModeChoice-MiD_k1234', d_biogeme_k1234)

# SAMPLE: city validation (Münster)
d_biogeme_outMS = pd.read_csv('F:/Processed Data/cityvalidation/ModeChoice_input_outMS.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_outMS = db.Database('ModeChoice-MiD_outMS', d_biogeme_outMS)

d_biogeme_inMS = pd.read_csv('F:/Processed Data/cityvalidation/ModeChoice_sim_inMS.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_inMS = db.Database('ModeChoice-MiD_inMS', d_biogeme_inMS)

# SAMPLE: city validation (Wuppertal)
d_biogeme_outWT = pd.read_csv('F:/Processed Data/cityvalidation/ModeChoice_input_outWT.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_outWT = db.Database('ModeChoice-MiD_outWT', d_biogeme_outWT)

d_biogeme_inWT = pd.read_csv('F:/Processed Data/cityvalidation/ModeChoice_sim_inWT.csv', usecols = ['choice','W_GEW_new', 'wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_inWT = db.Database('ModeChoice-MiD_inWT', d_biogeme_inWT)

# Define variables
weight = Variable('W_GEW_new')
weight_old = Variable('W_GEW')

choice = Variable('choice')

cbikeav = Variable('cbikeav')
ebikeav = Variable('ebikeav')
carav = Variable('P_STKFZ')

dist_ln = Variable('dist_ln')
#just for later analysis:
dist = Variable ('wegkm')

ticket = Variable('ticket')

spattyp_12 = Variable('spattyp_12')
spattyp_21 = Variable('spattyp_21')
spattyp_22 = Variable('spattyp_22')

grad = Variable('grad')
departures_ln = Variable('departures_ln')
cyclinfra = Variable('cyclinfra')

season_1 = Variable('season_1')
season_2 = Variable('season_2')
season_3 = Variable('season_3')
season_4 = Variable('season_4')

age_1 = Variable('age_1')
age_2 = Variable('age_2')
age_3 = Variable('age_3')
#age_4 only for interaction with slope
age_4 = Variable('age_4')
age_5 = Variable('age_5')
age_6 = Variable('age_6')
age_7 = Variable('age_7')
age_8 = Variable('age_8')

edu_1 = Variable('edu_1')
edu_2 = Variable('edu_2')
edu_3 = Variable('edu_3')
edu_5 = Variable('edu_5')
edu_6 = Variable('edu_6')

sex_2 = Variable('sex_2')

eco_1 = Variable('eco_1')
eco_2 = Variable('eco_2')
eco_4 = Variable('eco_4')
eco_5 = Variable('eco_5')

purp_2 = Variable('purp_2')
purp_3 = Variable('purp_3')
purp_4 = Variable('purp_4')
purp_5 = Variable('purp_5')
purp_6 = Variable('purp_6')
purp_7 = Variable('purp_7')

# Parameters to be estimated
# Arguments:
#   1  Name for report. Typically, the same as the variable
#   2  Starting value
#   3  Lower bound
#   4  Upper bound
#   5  0: estimate the parameter, 1: keep it fixed

#ASC
ASC_cbike = Beta('ASC_cbike', 0, None, None, 0)
ASC_ebike = Beta('ASC_ebike', 0, None, None, 0)
ASC_card = Beta('ASC_card', 0, None, None, 0)
ASC_carp = Beta('ASC_carp', 0, None, None, 0)
ASC_pt = Beta('ASC_pt', 0, None, None, 0)

#betas for trip distance
beta_dist_cbike = Beta('beta_dist_cbike', 0, None, None, 0)
beta_dist_ebike = Beta('beta_dist_ebike', 0, None, None, 0)
beta_dist_card = Beta('beta_dist_card', 0, None, None, 0)
beta_dist_carp = Beta('beta_dist_carp', 0, None, None, 0)
beta_dist_pt = Beta('beta_dist_pt', 0, None, None, 0)

#betas for pt season ticket
beta_ticket_pt = Beta('beta_ticket_pt', 0, None, None, 0)

#betas for slope
beta_grad_cbike_intage123 = Beta('beta_grad_cbike_intage123', 0, None, None, 0)
beta_grad_cbike_intage45 = Beta('beta_grad_cbike_intage45', 0, None, None, 0)
beta_grad_cbike_intage678 = Beta('beta_grad_cbike_intage678', 0, None, None, 0)

beta_grad_ebike_intage123 = Beta('beta_grad_ebike_intage123', 0, None, None, 0)
beta_grad_ebike_intage45 = Beta('beta_grad_ebike_intage45', 0, None, None, 0)
beta_grad_ebike_intage678 = Beta('beta_grad_ebike_intage678', 0, None, None, 0)

#betas for spatial typology
beta_spattyp_12_cbike = Beta('beta_spattyp_12_cbike', 0, None, None, 0)
beta_spattyp_21_cbike = Beta('beta_spattyp_21_cbike', 0, None, None, 0)
beta_spattyp_22_cbike = Beta('beta_spattyp_22_cbike', 0, None, None, 0)

beta_spattyp_12_ebike = Beta('beta_spattyp_12_ebike', 0, None, None, 0)
beta_spattyp_21_ebike = Beta('beta_spattyp_21_ebike', 0, None, None, 0)
beta_spattyp_22_ebike = Beta('beta_spattyp_22_ebike', 0, None, None, 0)

beta_spattyp_12_card = Beta('beta_spattyp_12_card', 0, None, None, 0)
beta_spattyp_21_card = Beta('beta_spattyp_21_card', 0, None, None, 0)
beta_spattyp_22_card = Beta('beta_spattyp_22_card', 0, None, None, 0)

beta_spattyp_12_carp = Beta('beta_spattyp_12_carp', 0, None, None, 0)
beta_spattyp_21_carp = Beta('beta_spattyp_21_carp', 0, None, None, 0)
beta_spattyp_22_carp = Beta('beta_spattyp_22_carp', 0, None, None, 0)

beta_spattyp_12_pt = Beta('beta_spattyp_12_pt', 0, None, None, 0)
beta_spattyp_21_pt = Beta('beta_spattyp_21_pt', 0, None, None, 0)
beta_spattyp_22_pt = Beta('beta_spattyp_22_pt', 0, None, None, 0)

#betas for cycling infrastructure
beta_cyclinfra_cbike = Beta('beta_cyclinfra_cbike', 0, None, None, 0)
beta_cyclinfra_ebike = Beta('beta_cyclinfra_ebike', 0, None, None, 0)

#betas for departures
beta_departures_ln_pt = Beta('beta_departures_ln_pt', 0, None, None, 0)

#betas for season
beta_season_1_cbike = Beta('beta_season_1_cbike', 0, None, None, 0)
beta_season_2_cbike = Beta('beta_season_2_cbike', 0, None, None, 0)
beta_season_4_cbike = Beta('beta_season_4_cbike', 0, None, None, 0)

beta_season_1_ebike = Beta('beta_season_1_ebike', 0, None, None, 0)
beta_season_2_ebike = Beta('beta_season_2_ebike', 0, None, None, 0)
beta_season_4_ebike = Beta('beta_season_4_ebike', 0, None, None, 0)

#betas for age
beta_age_1_cbike = Beta('beta_age_1_cbike', 0, None, None, 0)
beta_age_2_cbike = Beta('beta_age_2_cbike', 0, None, None, 0)
beta_age_3_cbike = Beta('beta_age_3_cbike', 0, None, None, 0)
beta_age_5_cbike = Beta('beta_age_5_cbike', 0, None, None, 0)
beta_age_6_cbike = Beta('beta_age_6_cbike', 0, None, None, 0)
beta_age_7_cbike = Beta('beta_age_7_cbike', 0, None, None, 0)
beta_age_8_cbike = Beta('beta_age_8_cbike', 0, None, None, 0)

beta_age_1_ebike = Beta('beta_age_1_ebike', 0, None, None, 0)
beta_age_2_ebike = Beta('beta_age_2_ebike', 0, None, None, 0)
beta_age_3_ebike = Beta('beta_age_3_ebike', 0, None, None, 0)
beta_age_5_ebike = Beta('beta_age_5_ebike', 0, None, None, 0)
beta_age_6_ebike = Beta('beta_age_6_ebike', 0, None, None, 0)
beta_age_7_ebike = Beta('beta_age_7_ebike', 0, None, None, 0)
beta_age_8_ebike = Beta('beta_age_8_ebike', 0, None, None, 0)

beta_age_1_card = Beta('beta_age_1_card', 0, None, None, 0)
beta_age_2_card = Beta('beta_age_2_card', 0, None, None, 0)
beta_age_3_card = Beta('beta_age_3_card', 0, None, None, 0)
beta_age_5_card = Beta('beta_age_5_card', 0, None, None, 0)
beta_age_6_card = Beta('beta_age_6_card', 0, None, None, 0)
beta_age_7_card = Beta('beta_age_7_card', 0, None, None, 0)
beta_age_8_card = Beta('beta_age_8_card', 0, None, None, 0)

beta_age_1_carp = Beta('beta_age_1_carp', 0, None, None, 0)
beta_age_2_carp = Beta('beta_age_2_carp', 0, None, None, 0)
beta_age_3_carp = Beta('beta_age_3_carp', 0, None, None, 0)
beta_age_5_carp = Beta('beta_age_5_carp', 0, None, None, 0)
beta_age_6_carp = Beta('beta_age_6_carp', 0, None, None, 0)
beta_age_7_carp = Beta('beta_age_7_carp', 0, None, None, 0)
beta_age_8_carp = Beta('beta_age_8_carp', 0, None, None, 0)

beta_age_1_pt = Beta('beta_age_1_pt', 0, None, None, 0)
beta_age_2_pt = Beta('beta_age_2_pt', 0, None, None, 0)
beta_age_3_pt = Beta('beta_age_3_pt', 0, None, None, 0)
beta_age_5_pt = Beta('beta_age_5_pt', 0, None, None, 0)
beta_age_6_pt = Beta('beta_age_6_pt', 0, None, None, 0)
beta_age_7_pt = Beta('beta_age_7_pt', 0, None, None, 0)
beta_age_8_pt = Beta('beta_age_8_pt', 0, None, None, 0)

#betas for edu
beta_edu_1_cbike = Beta('beta_edu_1_cbike', 0, None, None, 0)
beta_edu_2_cbike = Beta('beta_edu_2_cbike', 0, None, None, 0)
beta_edu_3_cbike = Beta('beta_edu_3_cbike', 0, None, None, 0)
beta_edu_5_cbike = Beta('beta_edu_5_cbike', 0, None, None, 0)
beta_edu_6_cbike = Beta('beta_edu_6_cbike', 0, None, None, 0)

beta_edu_1_ebike = Beta('beta_edu_1_ebike', 0, None, None, 0)
beta_edu_2_ebike = Beta('beta_edu_2_ebike', 0, None, None, 0)
beta_edu_3_ebike = Beta('beta_edu_3_ebike', 0, None, None, 0)
beta_edu_5_ebike = Beta('beta_edu_5_ebike', 0, None, None, 0)
beta_edu_6_ebike = Beta('beta_edu_6_ebike', 0, None, None, 0)

beta_edu_1_card = Beta('beta_edu_1_card', 0, None, None, 0)
beta_edu_2_card = Beta('beta_edu_2_card', 0, None, None, 0)
beta_edu_3_card = Beta('beta_edu_3_card', 0, None, None, 0)
beta_edu_5_card = Beta('beta_edu_5_card', 0, None, None, 0)
beta_edu_6_card = Beta('beta_edu_6_card', 0, None, None, 0)

beta_edu_1_carp = Beta('beta_edu_1_carp', 0, None, None, 0)
beta_edu_2_carp = Beta('beta_edu_2_carp', 0, None, None, 0)
beta_edu_3_carp = Beta('beta_edu_3_carp', 0, None, None, 0)
beta_edu_5_carp = Beta('beta_edu_5_carp', 0, None, None, 0)
beta_edu_6_carp = Beta('beta_edu_6_carp', 0, None, None, 0)

beta_edu_1_pt = Beta('beta_edu_1_pt', 0, None, None, 0)
beta_edu_2_pt = Beta('beta_edu_2_pt', 0, None, None, 0)
beta_edu_3_pt = Beta('beta_edu_3_pt', 0, None, None, 0)
beta_edu_5_pt = Beta('beta_edu_5_pt', 0, None, None, 0)
beta_edu_6_pt = Beta('beta_edu_6_pt', 0, None, None, 0)

#betas for sex
beta_sex_2_cbike = Beta('beta_sex_2_cbike', 0, None, None, 0)
beta_sex_2_ebike = Beta('beta_sex_2_ebike', 0, None, None, 0)
beta_sex_2_card = Beta('beta_sex_2_card', 0, None, None, 0)
beta_sex_2_carp = Beta('beta_sex_2_carp', 0, None, None, 0)
beta_sex_2_pt = Beta('beta_sex_2_pt', 0, None, None, 0)

#betas for economic status
beta_eco_1_cbike = Beta('beta_eco_1_cbike', 0, None, None, 0)
beta_eco_2_cbike = Beta('beta_eco_2_cbike', 0, None, None, 0)
beta_eco_4_cbike = Beta('beta_eco_4_cbike', 0, None, None, 0)
beta_eco_5_cbike = Beta('beta_eco_5_cbike', 0, None, None, 0)

beta_eco_1_ebike = Beta('beta_eco_1_ebike', 0, None, None, 0)
beta_eco_2_ebike = Beta('beta_eco_2_ebike', 0, None, None, 0)
beta_eco_4_ebike = Beta('beta_eco_4_ebike', 0, None, None, 0)
beta_eco_5_ebike = Beta('beta_eco_5_ebike', 0, None, None, 0)

beta_eco_1_card = Beta('beta_eco_1_card', 0, None, None, 0)
beta_eco_2_card = Beta('beta_eco_2_card', 0, None, None, 0)
beta_eco_4_card = Beta('beta_eco_4_card', 0, None, None, 0)
beta_eco_5_card = Beta('beta_eco_5_card', 0, None, None, 0)

beta_eco_1_carp = Beta('beta_eco_1_carp', 0, None, None, 0)
beta_eco_2_carp = Beta('beta_eco_2_carp', 0, None, None, 0)
beta_eco_4_carp = Beta('beta_eco_4_carp', 0, None, None, 0)
beta_eco_5_carp = Beta('beta_eco_5_carp', 0, None, None, 0)

beta_eco_1_pt = Beta('beta_eco_1_pt', 0, None, None, 0)
beta_eco_2_pt = Beta('beta_eco_2_pt', 0, None, None, 0)
beta_eco_4_pt = Beta('beta_eco_4_pt', 0, None, None, 0)
beta_eco_5_pt = Beta('beta_eco_5_pt', 0, None, None, 0)

#betas for trip purpose
beta_purp_2_cbike = Beta('beta_purp_2_cbike', 0, None, None, 0)
beta_purp_3_cbike = Beta('beta_purp_3_cbike', 0, None, None, 0)
beta_purp_4_cbike = Beta('beta_purp_4_cbike', 0, None, None, 0)
beta_purp_5_cbike = Beta('beta_purp_5_cbike', 0, None, None, 0)
beta_purp_6_cbike = Beta('beta_purp_6_cbike', 0, None, None, 0)
beta_purp_7_cbike = Beta('beta_purp_7_cbike', 0, None, None, 0)

beta_purp_2_ebike = Beta('beta_purp_2_ebike', 0, None, None, 0)
beta_purp_3_ebike = Beta('beta_purp_3_ebike', 0, None, None, 0)
beta_purp_4_ebike = Beta('beta_purp_4_ebike', 0, None, None, 0)
beta_purp_5_ebike = Beta('beta_purp_5_ebike', 0, None, None, 0)
beta_purp_6_ebike = Beta('beta_purp_6_ebike', 0, None, None, 0)
beta_purp_7_ebike = Beta('beta_purp_7_ebike', 0, None, None, 0)

beta_purp_2_card = Beta('beta_purp_2_card', 0, None, None, 0)
beta_purp_3_card = Beta('beta_purp_3_card', 0, None, None, 0)
beta_purp_4_card = Beta('beta_purp_4_card', 0, None, None, 0)
beta_purp_5_card = Beta('beta_purp_5_card', 0, None, None, 0)
beta_purp_6_card = Beta('beta_purp_6_card', 0, None, None, 0)
beta_purp_7_card = Beta('beta_purp_7_card', 0, None, None, 0)

beta_purp_2_carp = Beta('beta_purp_2_carp', 0, None, None, 0)
beta_purp_3_carp = Beta('beta_purp_3_carp', 0, None, None, 0)
beta_purp_4_carp = Beta('beta_purp_4_carp', 0, None, None, 0)
beta_purp_5_carp = Beta('beta_purp_5_carp', 0, None, None, 0)
beta_purp_6_carp = Beta('beta_purp_6_carp', 0, None, None, 0)
beta_purp_7_carp = Beta('beta_purp_7_carp', 0, None, None, 0)

beta_purp_2_pt = Beta('beta_purp_2_pt', 0, None, None, 0)
beta_purp_3_pt = Beta('beta_purp_3_pt', 0, None, None, 0)
beta_purp_4_pt = Beta('beta_purp_4_pt', 0, None, None, 0)
beta_purp_5_pt = Beta('beta_purp_5_pt', 0, None, None, 0)
beta_purp_6_pt = Beta('beta_purp_6_pt', 0, None, None, 0)
beta_purp_7_pt = Beta('beta_purp_7_pt', 0, None, None, 0)

# Utilities
foot = (
    0
)

cbike = (
    ASC_cbike -
    (age_1+age_2+age_3) * (grad**beta_grad_cbike_intage123) -
    (age_4+age_5) * (grad**beta_grad_cbike_intage45) -
    (age_6+age_7+age_8) * (grad**beta_grad_cbike_intage678) +
    beta_dist_cbike * dist_ln +
    beta_spattyp_12_cbike * spattyp_12 +
    beta_spattyp_21_cbike * spattyp_21 +
    beta_spattyp_22_cbike * spattyp_22 +
    beta_cyclinfra_cbike * cyclinfra +
    beta_season_1_cbike * season_1 +
    beta_season_2_cbike * season_2 +
    beta_season_4_cbike * season_4 +
    beta_age_1_cbike * age_1 +
    beta_age_2_cbike * age_2 +
    beta_age_3_cbike * age_3 +
    beta_age_5_cbike * age_5 +
    beta_age_6_cbike * age_6 +
    beta_age_7_cbike * age_7 +
    beta_age_8_cbike * age_8 +
    beta_edu_1_cbike * edu_1 +
    beta_edu_2_cbike * edu_2 +
    beta_edu_3_cbike * edu_3 +
    beta_edu_5_cbike * edu_5 +
    beta_sex_2_cbike * sex_2 +
    beta_eco_1_cbike * eco_1 +
    beta_eco_2_cbike * eco_2 +
    beta_eco_4_cbike * eco_4 +
    beta_eco_5_cbike * eco_5 +
    beta_purp_2_cbike * purp_2 +
    beta_purp_3_cbike * purp_3 +
    beta_purp_4_cbike * purp_4 +
    beta_purp_5_cbike * purp_5 +
    beta_purp_6_cbike * purp_6 +
    beta_purp_7_cbike * purp_7
)

ebike = (
    ASC_ebike -
    (age_1+age_2+age_3) * (grad**beta_grad_ebike_intage123) -
    (age_4+age_5) * (grad**beta_grad_ebike_intage45) -
    (age_6+age_7+age_8) * (grad**beta_grad_ebike_intage678) +
    beta_dist_ebike * dist_ln +
    beta_spattyp_12_ebike * spattyp_12 +
    beta_spattyp_21_ebike * spattyp_21 +
    beta_spattyp_22_ebike * spattyp_22 +
    beta_cyclinfra_ebike * cyclinfra +
    beta_season_1_ebike * season_1 +
    beta_season_2_ebike * season_2 +
    beta_season_4_ebike * season_4 +
    beta_age_1_ebike * age_1 +
    beta_age_2_ebike * age_2 +
    beta_age_3_ebike * age_3 +
    beta_age_5_ebike * age_5 +
    beta_age_6_ebike * age_6 +
    beta_age_7_ebike * age_7 +
    beta_age_8_ebike * age_8 +
    beta_edu_1_ebike * edu_1 +
    beta_edu_2_ebike * edu_2 +
    beta_edu_3_ebike * edu_3 +
    beta_edu_5_ebike * edu_5 +
    beta_sex_2_ebike * sex_2 +
    beta_eco_1_ebike * eco_1 +
    beta_eco_2_ebike * eco_2 +
    beta_eco_4_ebike * eco_4 +
    beta_eco_5_ebike * eco_5 +
    beta_purp_2_ebike * purp_2 +
    beta_purp_3_ebike * purp_3 +
    beta_purp_4_ebike * purp_4 +
    beta_purp_5_ebike * purp_5 +
    beta_purp_6_ebike * purp_6 +
    beta_purp_7_ebike * purp_7
)

card = (
    ASC_card +
    beta_dist_card * dist_ln +
    beta_spattyp_12_card * spattyp_12 +
    beta_spattyp_21_card * spattyp_21 +
    beta_spattyp_22_card * spattyp_22 +
    beta_age_1_card * age_1 +
    beta_age_2_card * age_2 +
    beta_age_3_card * age_3 +
    beta_age_5_card * age_5 +
    beta_age_6_card * age_6 +
    beta_age_7_card * age_7 +
    beta_age_8_card * age_8 +
    beta_edu_1_card * edu_1 +
    beta_edu_2_card * edu_2 +
    beta_edu_3_card * edu_3 +
    beta_edu_5_card * edu_5 +
    beta_sex_2_card * sex_2 +
    beta_eco_1_card * eco_1 +
    beta_eco_2_card * eco_2 +
    beta_eco_4_card * eco_4 +
    beta_eco_5_card * eco_5 +
    beta_purp_2_card * purp_2 +
    beta_purp_3_card * purp_3 +
    beta_purp_4_card * purp_4 +
    beta_purp_5_card * purp_5 +
    beta_purp_6_card * purp_6 +
    beta_purp_7_card * purp_7
)

carp = (
    ASC_carp +
    beta_dist_carp * dist_ln +
    beta_spattyp_12_carp * spattyp_12 +
    beta_spattyp_21_carp * spattyp_21 +
    beta_spattyp_22_carp * spattyp_22 +
    beta_age_1_carp * age_1 +
    beta_age_2_carp * age_2 +
    beta_age_3_carp * age_3 +
    beta_age_5_carp * age_5 +
    beta_age_6_carp * age_6 +
    beta_age_7_carp * age_7 +
    beta_age_8_carp * age_8 +
    beta_edu_1_carp * edu_1 +
    beta_edu_2_carp * edu_2 +
    beta_edu_3_carp * edu_3 +
    beta_edu_5_carp * edu_5 +
    beta_sex_2_carp * sex_2 +
    beta_eco_1_carp * eco_1 +
    beta_eco_2_carp * eco_2 +
    beta_eco_4_carp * eco_4 +
    beta_eco_5_carp * eco_5 +
    beta_purp_2_carp * purp_2 +
    beta_purp_3_carp * purp_3 +
    beta_purp_4_carp * purp_4 +
    beta_purp_5_carp * purp_5 +
    beta_purp_6_carp * purp_6 +
    beta_purp_7_carp * purp_7
)

pt = (
    ASC_pt +
    beta_dist_pt * dist_ln +
    beta_ticket_pt * ticket +
    beta_spattyp_12_pt * spattyp_12 +
    beta_spattyp_21_pt * spattyp_21 +
    beta_spattyp_22_pt * spattyp_22 +
    beta_departures_ln_pt * departures_ln +
    beta_age_1_pt * age_1 +
    beta_age_2_pt * age_2 +
    beta_age_3_pt * age_3 +
    beta_age_5_pt * age_5 +
    beta_age_6_pt * age_6 +
    beta_age_7_pt * age_7 +
    beta_age_8_pt * age_8 +
    beta_edu_1_pt * edu_1 +
    beta_edu_2_pt * edu_2 +
    beta_edu_3_pt * edu_3 +
    beta_edu_5_pt * edu_5 +
    beta_sex_2_pt * sex_2 +
    beta_eco_1_pt * eco_1 +
    beta_eco_2_pt * eco_2 +
    beta_eco_4_pt * eco_4 +
    beta_eco_5_pt * eco_5 +
    beta_purp_2_pt * purp_2 +
    beta_purp_3_pt * purp_3 +
    beta_purp_4_pt * purp_4 +
    beta_purp_5_pt * purp_5 +
    beta_purp_6_pt * purp_6 +
    beta_purp_7_pt * purp_7
)

V = {1: foot, 21: cbike, 22: ebike, 3: carp, 4: card, 5: pt}
av = {1: 1, 21: cbikeav, 22: ebikeav, 3: carav, 4: carav, 5: 1}

MU_car = Beta('MU_car', 2, 1, 10, 0)
car = OneNestForNestedLogit(
    nest_param=MU_car, list_of_alternatives=[3, 4], name='car'
)

nests = NestsForNestedLogit(choice_set=list(V), tuple_of_nests=(car,))

# ESTIMATE NESTED LOGIT MODEL (complete sample)
logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_complete, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_car'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

# ESTIMATE NESTED LOGIT MODEL (k-fold cross validation)
logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_k2345, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_k2345'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_k1345, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_k1345'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_k1245, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_k1245'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_k1235, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_k1235'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_k1234, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_k1234'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

# ESTIMATE LOGIT MODEL (cityvalidation MS)
logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_outMS, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_outMS'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

# ESTIMATE LOGIT MODEL (cityvalidation WT)
logprob = models.lognested(V, av, nests, choice)
bio.weight = weight
biogeme = bio.BIOGEME(database_outWT, logprob)
biogeme.modelName = 'ModeChoice-MiD_NL_outWT'
biogeme.calculate_null_loglikelihood(av)
results = biogeme.estimate()

#######################################
#  #  #   V A L I D A T I O N   #  #  #
#######################################

import biogeme.results as res
import biogeme.database as db
import biogeme.biogeme as bio
from biogeme.models import logit
from biogeme.expressions import Variable, Derive
from biogeme import models

prob_Opt1 = models.nested(V, av, nests, 1)
prob_Opt21 = models.nested(V, av, nests, 21)
prob_Opt22 = models.nested(V, av, nests, 22)
prob_Opt3 = models.nested(V, av, nests, 3)
prob_Opt4 = models.nested(V, av, nests, 4)
prob_Opt5 = models.nested(V, av, nests, 5)

triplength = Variable('wegkm')
age = Variable('age')
purp = Variable('purp')

simulate = {
    'foot_prob': prob_Opt1,
    'cbike_prob': prob_Opt21,
    'ebike_prob': prob_Opt22,
    'carp_prob': prob_Opt3,
    'card_prob': prob_Opt4,
    'pt_prob': prob_Opt5,
    'triplength': triplength,
    'age': age,
    'purp': purp,
    'choice': choice
}
# in simulation "triplength", "purp" and "age" is included to allow for more in-depth analysis. The variable "choice" is copied over on the simulation results to allow for comparison between real and simulated choice.

def determine_sim_choice(row):
    if row['foot_prob'] > row[['cbike_prob', 'ebike_prob', 'carp_prob', 'card_prob', 'pt_prob']].max():
        return 1
    elif row['cbike_prob'] == row[['cbike_prob', 'ebike_prob', 'carp_prob', 'card_prob', 'pt_prob']].max():
        return 21
    elif row['ebike_prob'] == row[['cbike_prob', 'ebike_prob', 'carp_prob', 'card_prob', 'pt_prob']].max():
        return 22
    elif row['carp_prob'] == row[['cbike_prob', 'ebike_prob', 'carp_prob', 'card_prob', 'pt_prob']].max():
        return 3
    elif row['card_prob'] == row[['cbike_prob', 'ebike_prob', 'carp_prob', 'card_prob', 'pt_prob']].max():
        return 4
    else:
        return 5

# # #   S I M U L A T I O N   # # #

#k-fold cross validation - simulation
#use model estimated on four out of five batched of trips to simulate choice for the fifth batch

#use model estimated on batches 1,2,3 and 4 to simulate batch 5
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_k1234.pickle')

d_biogeme_sim_k5 = pd.read_csv('D:/Processed Data/kfcv/ModeChoice_sim_k5.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_k5 = db.Database('database_sim_k5', d_biogeme_sim_k5)

biosim = bio.BIOGEME(database_sim_k5, simulate)
biosim.modelName = 'sim_k5'
simresults_k5 = biosim.simulate(results.get_beta_values())

simresults_k5['choice_sim'] = simresults_k5.apply(determine_sim_choice, axis=1)

simresults_k5[['age']] = simresults_k5[['age']].astype(int)
simresults_k5[['purp']] = simresults_k5[['purp']].astype(int)
simresults_k5[['choice']]= simresults_k5[['choice']].astype(int)

simresults_k5.to_csv('D:/Processed Data/kfcv/simresults/simresults_k5.csv')
#simresults_k5

#use model estimated on batches 1,2,3 and 5 to simulate batch 4
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_k1235.pickle')

d_biogeme_sim_k4 = pd.read_csv('D:/Processed Data/kfcv/ModeChoice_sim_k4.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_k4 = db.Database('database_sim_k4', d_biogeme_sim_k4)

biosim = bio.BIOGEME(database_sim_k4, simulate)
biosim.modelName = 'sim_k4'
simresults_k4 = biosim.simulate(results.get_beta_values())

simresults_k4['choice_sim'] = simresults_k4.apply(determine_sim_choice, axis=1)

simresults_k4[['age']] = simresults_k4[['age']].astype(int)
simresults_k4[['purp']] = simresults_k4[['purp']].astype(int)
simresults_k4[['choice']]= simresults_k4[['choice']].astype(int)

simresults_k4.to_csv('D:/Processed Data/kfcv/simresults/simresults_k4.csv')
#simresults_k4

#use model estimated on batches 1,2,4 and 5 to simulate batch 3
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_k1245.pickle')

d_biogeme_sim_k3 = pd.read_csv('D:/Processed Data/kfcv/ModeChoice_sim_k3.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_k3 = db.Database('database_sim_k3', d_biogeme_sim_k3)

biosim = bio.BIOGEME(database_sim_k3, simulate)
biosim.modelName = 'sim_k3'
simresults_k3 = biosim.simulate(results.get_beta_values())

simresults_k3['choice_sim'] = simresults_k3.apply(determine_sim_choice, axis=1)

simresults_k3[['age']] = simresults_k3[['age']].astype(int)
simresults_k3[['purp']] = simresults_k3[['purp']].astype(int)
simresults_k3[['choice']]= simresults_k3[['choice']].astype(int)

simresults_k3.to_csv('D:/Processed Data/kfcv/simresults/simresults_k3.csv')
#simresults_k3

#use model estimated on batches 1,3,4 and 5 to simulate batch 2
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_k1345.pickle')

d_biogeme_sim_k2 = pd.read_csv('D:/Processed Data/kfcv/ModeChoice_sim_k2.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_k2 = db.Database('database_sim_k2', d_biogeme_sim_k2)

biosim = bio.BIOGEME(database_sim_k2, simulate)
biosim.modelName = 'sim_k2'
simresults_k2 = biosim.simulate(results.get_beta_values())

simresults_k2['choice_sim'] = simresults_k2.apply(determine_sim_choice, axis=1)

simresults_k2[['age']] = simresults_k2[['age']].astype(int)
simresults_k2[['purp']] = simresults_k2[['purp']].astype(int)
simresults_k2[['choice']]= simresults_k2[['choice']].astype(int)

simresults_k2.to_csv('D:/Processed Data/kfcv/simresults/simresults_k2.csv')
#simresults_k2

#use model estimated on batches 2,3,4 and 5 to simulate batch 1
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_k2345.pickle')

d_biogeme_sim_k1 = pd.read_csv('D:/Processed Data/kfcv/ModeChoice_sim_k1.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_k1 = db.Database('database_sim_k1', d_biogeme_sim_k1)

biosim = bio.BIOGEME(database_sim_k1, simulate)
biosim.modelName = 'sim_k1'
simresults_k1 = biosim.simulate(results.get_beta_values())

simresults_k1['choice_sim'] = simresults_k1.apply(determine_sim_choice, axis=1)

simresults_k1[['age']] = simresults_k1[['age']].astype(int)
simresults_k1[['purp']] = simresults_k1[['purp']].astype(int)
simresults_k1[['choice']]= simresults_k1[['choice']].astype(int)

simresults_k1.to_csv('D:/Processed Data/kfcv/simresults/simresults_k1.csv')
#simresults_k1

#city-validation - simulation
#use model estimated on all trips from people living outside of city X to simulate all trips undertaken by people living inside city X

#use model estimated on trips by people living outside of Münster (NRW) to simulate trips undertaken by people living inside Münster
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_outMS.pickle')

d_biogeme_sim_MS = pd.read_csv('D:/Processed Data/cityvalidation/ModeChoice_sim_inMS.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_MS = db.Database('database_sim_MS', d_biogeme_sim_MS)

biosim = bio.BIOGEME(database_sim_MS, simulate)
biosim.modelName = 'sim_MS'
simresults_MS = biosim.simulate(results.getBetaValues())

simresults_MS['choice_sim'] = simresults_MS.apply(determine_sim_choice, axis=1)
simresults_MS[['age']] = simresults_MS[['age']].astype(int)
simresults_MS[['purp']] = simresults_MS[['purp']].astype(int)
simresults_MS[['choice']]= simresults_MS[['choice']].astype(int)

simresults_MS.to_csv('D:/Processed Data/cityvalidation/simresults/simresults_MS.csv')
simresults_MS

#use model estimated on trips by people living outside of Wuppertal (NRW) to simulate trips undertaken by people living inside Wuppertal
results = res.bioResults(pickle_file='Estimation results/Final/ModeChoice-MiD_NL_outWT.pickle')

d_biogeme_sim_WT = pd.read_csv('D:/Processed Data/cityvalidation/ModeChoice_sim_inWT.csv',usecols = ['choice','W_GEW_new','age','purp','wegkm','dist_ln','cbikeav','ebikeav','P_STKFZ','ticket','age_1','age_2','age_3','age_4','age_5','age_6','age_7','age_8','sex_2','edu_1','edu_2','edu_3','edu_5','edu_6','eco_1','eco_2','eco_4','eco_5','purp_2','purp_3','purp_4','purp_5','purp_6','purp_7','season_1','season_2','season_4','spattyp_12','spattyp_21','spattyp_22','grad','departures_ln','departures_ln','cyclinfra'])
database_sim_WT = db.Database('database_sim_WT', d_biogeme_sim_WT)

biosim = bio.BIOGEME(database_sim_WT, simulate)
biosim.modelName = 'sim_WT'
simresults_WT = biosim.simulate(results.get_beta_values())

simresults_WT['choice_sim'] = simresults_WT.apply(determine_sim_choice, axis=1)

simresults_WT[['age']] = simresults_WT[['age']].astype(int)
simresults_WT[['purp']] = simresults_WT[['purp']].astype(int)
simresults_WT[['choice']]= simresults_WT[['choice']].astype(int)

simresults_WT.to_csv('D:/Processed Data/cityvalidation/simresults/simresults_WT.csv')
simresults_WT

##########################
### SUBSTITUTION RATES ###
##########################

# Method 1
# uses NL model estimated on complete sample
# take all trips that were undertaken on an e-bike, change e-bike av. to 0, then simulate new mode choice

# re-define "simulate" from earlier, since age, purpose and triplength are not needed for this simulation
simulate = {
    'foot_prob': prob_Opt1,
    'cbike_prob': prob_Opt21,
    'ebike_prob': prob_Opt22,
    'carp_prob': prob_Opt3,
    'card_prob': prob_Opt4,
    'pt_prob': prob_Opt5,
    'triplength': triplength,
    'weight_old': weight_old,
}

results = res.bioResults(pickle_file='Estimation results/Final//ModeChoice-MiD_NL_car.pickle')

d_biogeme_sim_substitution = d_biogeme_complete[d_biogeme_complete['choice'] == 22]

d_biogeme_sim_substitution.loc[:, 'ebikeav'] = 0
database_sim_substitution = db.Database('database_sim_substitution', d_biogeme_sim_substitution)

biosim = bio.BIOGEME(database_sim_substitution, simulate)
biosim.modelName = 'sim_substitution'
simresults_simulation = biosim.simulate(results.getBetaValues())
simresults_simulation.to_csv('D:/Processed Data/simresults_substitution_m1.csv')
simresults_simulation

# Method 2
# take all trips that were undertaken on an e-bike, change e-bike av. to 0, AND C-BIKE AV. TO 1, then simulate new mode choice

# re-define "simulate" from earlier, since age, purpose and triplength are not needed for this simulation
simulate = {
    'foot_prob': prob_Opt1,
    'cbike_prob': prob_Opt21,
    'ebike_prob': prob_Opt22,
    'carp_prob': prob_Opt3,
    'card_prob': prob_Opt4,
    'pt_prob': prob_Opt5,
    'triplength': triplength,
    'weight_old': weight_old,
}

results = res.bioResults(pickle_file='Estimation results/Final//ModeChoice-MiD_NL_car.pickle')

d_biogeme_sim_substitution = d_biogeme_complete[d_biogeme_complete['choice'] == 22]

d_biogeme_sim_substitution.loc[:, 'ebikeav'] = 0
d_biogeme_sim_substitution.loc[:, 'cbikeav'] = 1

database_sim_substitution = db.Database('database_sim_substitution', d_biogeme_sim_substitution)

biosim = bio.BIOGEME(database_sim_substitution, simulate)
biosim.modelName = 'full_sim_substitution_m2'
simresults_simulation_m2 = biosim.simulate(results.getBetaValues())
simresults_simulation_m2.to_csv('D:/Processed Data/simresults_substitution_m2.csv')
simresults_simulation_m2

# ELASTICITIES

results = res.bioResults(pickle_file='Estimation results/Final//ModeChoice-MiD_NL_car.pickle')

# point elasticity for change in infra and impact on cbike
point_elas_cbike_infra = Derive(prob_Opt21, 'cyclinfra') * cyclinfra / prob_Opt21

simulate = {
    'weight': weight,
    'cbike_prob': prob_Opt21,
    'Weighted cbike_prob': prob_Opt21 * weight,
    'Point elast. cbike infra': point_elas_cbike_infra,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. cbike infra'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_cbike_infra = (
    simresults['Weighted cbike_prob']
    * simresults['Point elast. cbike infra']
    / simresults['Weighted cbike_prob'].sum()
    ).sum()
    
aggregate_point_elas_cbike_infra

# point elasticity for change in infra and impact on ebike
point_elas_ebike_infra = Derive(prob_Opt22, 'cyclinfra') * cyclinfra / prob_Opt22

simulate = {
    'weight': weight,
    'ebike_prob': prob_Opt22,
    'Weighted ebike_prob': prob_Opt22 * weight,
    'Point elast. ebike infra': point_elas_ebike_infra,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. ebike infra'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_ebike_infra = (
    simresults['Weighted ebike_prob']
    * simresults['Point elast. ebike infra']
    / simresults['Weighted ebike_prob'].sum()
    ).sum()
    
aggregate_point_elas_ebike_infra

# point elasticity for change in dist_ln and impact on cbike
point_elas_cbike_dist_ln = Derive(prob_Opt21, 'dist_ln') * dist_ln / prob_Opt21

simulate = {
    'weight': weight,
    'cbike_prob': prob_Opt21,
    'Weighted cbike_prob': prob_Opt21 * weight,
    'Point elast. cbike dist_ln': point_elas_cbike_dist_ln,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. cbike dist_ln'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_cbike_dist_ln = (
    simresults['Weighted cbike_prob']
    * simresults['Point elast. cbike dist_ln']
    / simresults['Weighted cbike_prob'].sum()
    ).sum()
    
aggregate_point_elas_cbike_dist_ln

# point elasticity for change in dist_ln and impact on ebike
point_elas_ebike_dist_ln = Derive(prob_Opt22, 'dist_ln') * dist_ln / prob_Opt22

simulate = {
    'weight': weight,
    'ebike_prob': prob_Opt22,
    'Weighted ebike_prob': prob_Opt22 * weight,
    'Point elast. ebike dist_ln': point_elas_ebike_dist_ln,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. ebike dist_ln'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_ebike_dist_ln = (
    simresults['Weighted ebike_prob']
    * simresults['Point elast. ebike dist_ln']
    / simresults['Weighted ebike_prob'].sum()
    ).sum()
    
aggregate_point_elas_ebike_dist_ln

# point elasticity for change in gradient and impact on cbike
point_elas_cbike_grad = Derive(prob_Opt21, 'grad') * grad / prob_Opt21

simulate = {
    'weight': weight,
    'cbike_prob': prob_Opt21,
    'Weighted cbike_prob': prob_Opt21 * weight,
    'Point elast. cbike grad': point_elas_cbike_grad,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. cbike grad'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_cbike_grad = (
    simresults['Weighted cbike_prob']
    * simresults['Point elast. cbike grad']
    / simresults['Weighted cbike_prob'].sum()
    ).sum()
    
aggregate_point_elas_cbike_grad

# point elasticity for change in grad and impact on ebike
point_elas_ebike_grad = Derive(prob_Opt22, 'grad') * grad / prob_Opt22

simulate = {
    'weight': weight,
    'ebike_prob': prob_Opt22,
    'Weighted ebike_prob': prob_Opt22 * weight,
    'Point elast. ebike grad': point_elas_ebike_grad,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'point elas. ebike grad'
simresults = biosim.simulate(results.get_beta_values())

aggregate_point_elas_ebike_grad = (
    simresults['Weighted ebike_prob']
    * simresults['Point elast. ebike grad']
    / simresults['Weighted ebike_prob'].sum()
    ).sum()
    
aggregate_point_elas_ebike_grad

# cross point elasticity for change in pt departures and impact on cbike
cross_point_elas_cbike_departures_ln = Derive(prob_Opt21, 'departures_ln') * departures_ln / prob_Opt21

simulate = {
    'weight': weight,
    'cbike_prob': prob_Opt21,
    'Weighted cbike_prob': prob_Opt21 * weight,
    'Cross point elast. cbike departures_ln': cross_point_elas_cbike_departures_ln,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'cross point elas. cbike departures_ln'
simresults = biosim.simulate(results.get_beta_values())

aggregate_cross_point_elas_cbike_departures_ln = (
    simresults['Weighted cbike_prob']
    * simresults['Cross point elast. cbike departures_ln']
    / simresults['Weighted cbike_prob'].sum()
    ).sum()
    
aggregate_cross_point_elas_cbike_departures_ln

# cross point elasticity for change in pt departures at trip origin and impact on ebike
cross_point_elas_ebike_departures_ln = Derive(prob_Opt22, 'departures_ln') * departures_ln / prob_Opt22

simulate = {
    'weight': weight,
    'ebike_prob': prob_Opt22,
    'Weighted ebike_prob': prob_Opt22 * weight,
    'Cross point elast. ebike departures_ln': cross_point_elas_ebike_departures_ln,
}

biosim = bio.BIOGEME(database_complete, simulate)
biosim.modelName = 'cross point elas. ebike departures_ln'
simresults = biosim.simulate(results.get_beta_values())

aggregate_cross_point_elas_ebike_departures_ln = (
    simresults['Weighted ebike_prob']
    * simresults['Cross point elast. ebike departures_ln']
    / simresults['Weighted ebike_prob'].sum()
    ).sum()
    
aggregate_cross_point_elas_ebike_departures_ln
