==> 30_Analysis_Validation.ipynb <==
import pandas as pd
import numpy as np

simresults_k1 = pd.read_csv('D:/Processed Data/kfcv/simresults/simresults_k1.csv')
simresults_k2 = pd.read_csv('D:/Processed Data/kfcv/simresults/simresults_k2.csv')
simresults_k3 = pd.read_csv('D:/Processed Data/kfcv/simresults/simresults_k3.csv')
simresults_k4 = pd.read_csv('D:/Processed Data/kfcv/simresults/simresults_k4.csv')
simresults_k5 = pd.read_csv('D:/Processed Data/kfcv/simresults/simresults_k5.csv')
simresults_k12345 = pd.concat([simresults_k1, simresults_k2, simresults_k3, simresults_k4, simresults_k5], sort=False)

simresults_MS = pd.read_csv('D:/Processed Data/cityvalidation/simresults/simresults_MS.csv')
simresults_WT = pd.read_csv('D:/Processed Data/cityvalidation/simresults/simresults_WT.csv')

simresults_substitution_m1 = pd.read_csv('D:/Processed Data/simresults_substitution_m1.csv')
simresults_substitution_m2 = pd.read_csv('D:/Processed Data/simresults_substitution_m2.csv')

###################
### MODE SHARES ###
###################

#mode shares are calculated without taking into account the weights for each trip. 

all_choices = pd.Series([1, 3, 4, 5, 21, 22])
columns_to_sum = ['foot_prob', 'carp_prob', 'card_prob', 'pt_prob', 'cbike_prob', 'ebike_prob']
mode_reweighting = pd.read_csv('../mode-reweighting.csv')
mode_reweighting

#kfcv, mode share for all trips

# mode share in the data
modeshare_recorded = simresults_k12345['choice'].value_counts() / len(simresults_k12345)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345['choice_sim'].value_counts() / len(simresults_k12345)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345[column]) / len(simresults_k12345)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)

overview_unweighted.to_csv('../validation/modeshares_kfcv_alltrips.csv')
overview_unweighted

#kfcv, mode share by distance class (0-5, 5-10, 10-50, 50-100 km)

#dividing the dataset into four distance classes
simresults_k12345_dist1 = simresults_k12345[simresults_k12345['triplength'] < 5]
simresults_k12345_dist2 = simresults_k12345[(simresults_k12345['triplength'] >= 5) & (simresults_k12345['triplength'] < 10)]
simresults_k12345_dist3 = simresults_k12345[(simresults_k12345['triplength'] >= 10) & (simresults_k12345['triplength'] < 50)]
simresults_k12345_dist4 = simresults_k12345[simresults_k12345['triplength'] >= 50]

#kfcv, distance class 1

# mode share in the data
modeshare_recorded = simresults_k12345_dist1['choice'].value_counts() / len(simresults_k12345_dist1)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_dist1['choice_sim'].value_counts() / len(simresults_k12345_dist1)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_dist1[column]) / len(simresults_k12345_dist1)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_dist1.csv')
overview_unweighted

#kfcv, distance class 2

# mode share in the data
modeshare_recorded = simresults_k12345_dist2['choice'].value_counts() / len(simresults_k12345_dist2)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_dist2['choice_sim'].value_counts() / len(simresults_k12345_dist2)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_dist2[column]) / len(simresults_k12345_dist2)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_dist2.csv')
overview_unweighted

#kfcv, distance class 3

# mode share in the data
modeshare_recorded = simresults_k12345_dist3['choice'].value_counts() / len(simresults_k12345_dist3)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_dist3['choice_sim'].value_counts() / len(simresults_k12345_dist3)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_dist3[column]) / len(simresults_k12345_dist3)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_dist3.csv')
overview_unweighted

#kfcv, distance class 4

# mode share in the data
modeshare_recorded = simresults_k12345_dist4['choice'].value_counts() / len(simresults_k12345_dist4)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_dist4['choice_sim'].value_counts() / len(simresults_k12345_dist4)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_dist4[column]) / len(simresults_k12345_dist4)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_dist4.csv')
overview_unweighted

#kfcv, mode share by age class

#dividing the dataset into eight age classes
#simresults_k12345_age1 = simresults_k12345[simresults_k12345['age'] ==1]
#simresults_k12345_age2 = simresults_k12345[simresults_k12345['age'] ==2]
#simresults_k12345_age3 = simresults_k12345[simresults_k12345['age'] ==3]
#simresults_k12345_age4 = simresults_k12345[simresults_k12345['age'] ==4]
#simresults_k12345_age5 = simresults_k12345[simresults_k12345['age'] ==5]
#simresults_k12345_age6 = simresults_k12345[simresults_k12345['age'] ==6]
#simresults_k12345_age7 = simresults_k12345[simresults_k12345['age'] ==7]
#simresults_k12345_age8 = simresults_k12345[simresults_k12345['age'] ==8]

#dividing the dataset into four age classes
simresults_k12345_age12 = simresults_k12345[(simresults_k12345['age'] == 1) | (simresults_k12345['age'] == 2)]
simresults_k12345_age34 = simresults_k12345[(simresults_k12345['age'] == 3) | (simresults_k12345['age'] == 4)]
simresults_k12345_age56 = simresults_k12345[(simresults_k12345['age'] == 5) | (simresults_k12345['age'] == 6)]
simresults_k12345_age78 = simresults_k12345[(simresults_k12345['age'] == 7) | (simresults_k12345['age'] == 8)]

#kfc, age class 12

# mode share in the data
modeshare_recorded = simresults_k12345_age12['choice'].value_counts() / len(simresults_k12345_age12)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_age12['choice_sim'].value_counts() / len(simresults_k12345_age12)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_age12[column]) / len(simresults_k12345_age12)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_age12.csv')
overview_unweighted

#kfc, age class 34

# mode share in the data
modeshare_recorded = simresults_k12345_age34['choice'].value_counts() / len(simresults_k12345_age34)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_age34['choice_sim'].value_counts() / len(simresults_k12345_age34)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_age34[column]) / len(simresults_k12345_age34)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_age34.csv')
overview_unweighted

#kfc, age class 56

# mode share in the data
modeshare_recorded = simresults_k12345_age56['choice'].value_counts() / len(simresults_k12345_age56)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_age56['choice_sim'].value_counts() / len(simresults_k12345_age56)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_age56[column]) / len(simresults_k12345_age56)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_age56.csv')
overview_unweighted

#kfc, age class 78

# mode share in the data
modeshare_recorded = simresults_k12345_age78['choice'].value_counts() / len(simresults_k12345_age78)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_age78['choice_sim'].value_counts() / len(simresults_k12345_age78)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_age78[column]) / len(simresults_k12345_age78)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_age78.csv')
overview_unweighted

#kfcv, mode share by trip purpose

#dividing the dataset into four distance classes
simresults_k12345_purp1 = simresults_k12345[simresults_k12345['purp'] ==1]
simresults_k12345_purp2 = simresults_k12345[simresults_k12345['purp'] ==2]
simresults_k12345_purp3 = simresults_k12345[simresults_k12345['purp'] ==3]
simresults_k12345_purp4 = simresults_k12345[simresults_k12345['purp'] ==4]
simresults_k12345_purp5 = simresults_k12345[simresults_k12345['purp'] ==5]
simresults_k12345_purp6 = simresults_k12345[simresults_k12345['purp'] ==6]
simresults_k12345_purp7 = simresults_k12345[simresults_k12345['purp'] ==7]

#kfcv, purpose 1

# mode share in the data
modeshare_recorded = simresults_k12345_purp1['choice'].value_counts() / len(simresults_k12345_purp1)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp1['choice_sim'].value_counts() / len(simresults_k12345_purp1)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp1[column]) / len(simresults_k12345_purp1)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp1.csv')
overview_unweighted

#kfcv, purpose 2

# mode share in the data
modeshare_recorded = simresults_k12345_purp2['choice'].value_counts() / len(simresults_k12345_purp2)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp2['choice_sim'].value_counts() / len(simresults_k12345_purp2)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp2[column]) / len(simresults_k12345_purp2)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp2.csv')
overview_unweighted

#kfcv, purpose 3

# mode share in the data
modeshare_recorded = simresults_k12345_purp3['choice'].value_counts() / len(simresults_k12345_purp3)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp3['choice_sim'].value_counts() / len(simresults_k12345_purp3)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp3[column]) / len(simresults_k12345_purp3)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp3.csv')
overview_unweighted

#kfcv, purpose 4

# mode share in the data
modeshare_recorded = simresults_k12345_purp4['choice'].value_counts() / len(simresults_k12345_purp4)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp4['choice_sim'].value_counts() / len(simresults_k12345_purp4)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp4[column]) / len(simresults_k12345_purp4)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp4.csv')
overview_unweighted

#kfcv, purpose 5

# mode share in the data
modeshare_recorded = simresults_k12345_purp5['choice'].value_counts() / len(simresults_k12345_purp5)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp5['choice_sim'].value_counts() / len(simresults_k12345_purp5)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp5[column]) / len(simresults_k12345_purp5)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp5.csv')
overview_unweighted

#kfcv, purpose 6

# mode share in the data
modeshare_recorded = simresults_k12345_purp6['choice'].value_counts() / len(simresults_k12345_purp6)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp6['choice_sim'].value_counts() / len(simresults_k12345_purp6)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp6[column]) / len(simresults_k12345_purp6)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp6.csv')
overview_unweighted

#kfcv, purpose 7

# mode share in the data
modeshare_recorded = simresults_k12345_purp7['choice'].value_counts() / len(simresults_k12345_purp7)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_k12345_purp7['choice_sim'].value_counts() / len(simresults_k12345_purp7)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_k12345_purp7[column]) / len(simresults_k12345_purp7)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_kfcv_purp7.csv')
overview_unweighted

#city-validation, Münster

# mode share in the data
modeshare_recorded = simresults_MS['choice'].value_counts() / len(simresults_MS)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_MS['choice_sim'].value_counts() / len(simresults_MS)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_MS[column]) / len(simresults_MS)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_MS.csv')
overview_unweighted

#city-validation, Wuppertal

# mode share in the data
modeshare_recorded = simresults_WT['choice'].value_counts() / len(simresults_WT)
modeshare_recorded = modeshare_recorded.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking most likely mode for each trip
modeshare_sim_ml = simresults_WT['choice_sim'].value_counts() / len(simresults_WT)
modeshare_sim_ml = modeshare_sim_ml.reindex(all_choices, fill_value=0)

# mode share in the simulation results, taking probabilities each trip
modeshare_sim_prob = []

for column in columns_to_sum:
    prob = np.sum(simresults_WT[column]) / len(simresults_WT)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'recorded': [modeshare_recorded[1],modeshare_recorded[3],modeshare_recorded[4],modeshare_recorded[5],modeshare_recorded[21],modeshare_recorded[22]],
     'sim_mostlikely': [modeshare_sim_ml[1],modeshare_sim_ml[3],modeshare_sim_ml[4],modeshare_sim_ml[5],modeshare_sim_ml[21],modeshare_sim_ml[22]],
    'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)
overview_unweighted.to_csv('../validation/modeshares_WT.csv')
overview_unweighted

##########################
### SUBSTITUTION RATES ###
##########################

simresults_substitution_m1 = pd.read_csv('D:/Processed Data/simresults_substitution_m1.csv')
simresults_substitution_m2 = pd.read_csv('D:/Processed Data/simresults_substitution_m2.csv')

# Method 1
#trip based substitution

modeshare_sim_prob = []
weights = simresults_substitution_m1["weight_old"]
for column in columns_to_sum:
    
    prob = np.sum(simresults_substitution_m1[column] * weights) / np.sum(weights)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)

overview_unweighted.to_csv('../validation/substitution_ebike_trips_m1.csv')

overview_unweighted

# Method 1
# trip length based substitution

modeshare_sim_prob_length = []
weights = simresults_substitution_m1["weight_old"]
weighted_triplength = simresults_substitution_m1["triplength"] * weights

for column in columns_to_sum:
    prob = np.sum(simresults_substitution_m1[column] * weighted_triplength) / np.sum(weighted_triplength)
    modeshare_sim_prob_length.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'sim_prob_length': [modeshare_sim_prob_length[0],modeshare_sim_prob_length[1],modeshare_sim_prob_length[2],modeshare_sim_prob_length[3],modeshare_sim_prob_length[4],modeshare_sim_prob_length[5]]}

overview_unweighted = pd.DataFrame(data=d)

overview_unweighted.to_csv('../validation/substitution_ebike_dist_m1.csv')

overview_unweighted

# Method 2
#trip based substitution

modeshare_sim_prob = []
weights = simresults_substitution_m2["weight_old"]

for column in columns_to_sum:
    prob = np.sum(simresults_substitution_m2[column] * weights) / np.sum(weights)
    modeshare_sim_prob.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'sim_prob': [modeshare_sim_prob[0],modeshare_sim_prob[1],modeshare_sim_prob[2],modeshare_sim_prob[3],modeshare_sim_prob[4],modeshare_sim_prob[5]]}

overview_unweighted = pd.DataFrame(data=d)

overview_unweighted.to_csv('../validation/substitution_ebike_trips_m2.csv')

overview_unweighted

# Method 2
#length based substitution

modeshare_sim_prob_length = []
weights = simresults_substitution_m2["weight_old"]
weighted_triplength = simresults_substitution_m2["triplength"] * weights

for column in columns_to_sum:
    prob = np.sum(simresults_substitution_m2[column] * weighted_triplength) / np.sum(weighted_triplength)
    modeshare_sim_prob_length.append(prob)
    
d = {'mode': ['foot','carp','card','pt','cbike','ebike'], 'sim_prob_length': [modeshare_sim_prob_length[0],modeshare_sim_prob_length[1],modeshare_sim_prob_length[2],modeshare_sim_prob_length[3],modeshare_sim_prob_length[4],modeshare_sim_prob_length[5]]}

overview_unweighted = pd.DataFrame(data=d)

overview_unweighted.to_csv('../validation/substitution_ebike_dist_m2.csv')

overview_unweighted

# The results above need to be readjusted with the mode-level reweighting factors, since the model is estimated on a sample that undersampled bike and walking trips
reweight = pd.read_csv('C:/Users/arning.FUL/Desktop/11_Mode_Choice/mode-reweighting.csv')

sub_ebike_trips_m1 = pd.read_csv('../validation/substitution_ebike_trips_m1.csv')
sub_ebike_dist_m1 = pd.read_csv('../validation/substitution_ebike_dist_m1.csv')
sub_ebike_trips_m2 = pd.read_csv('../validation/substitution_ebike_trips_m2.csv')
sub_ebike_dist_m2 = pd.read_csv('../validation/substitution_ebike_dist_m2.csv')

mode_mapping = {
    'foot': 1,
    'carp': 3,
    'card': 4,
    'pt': 5,
    'cbike': 21,
    'ebike': 22
}

reweight
#sub_ebike_trips_m1

# Reweight trip substitution, m1
sub_ebike_trips_m1['mode_int'] = sub_ebike_trips_m1['mode'].map(mode_mapping)
merged_df = pd.merge(sub_ebike_trips_m1, reweight, left_on='mode_int', right_on='mode', how='left')
merged_df['sim_prob_rew'] = merged_df['sim_prob'] * merged_df['reweight']
total_prob = merged_df['sim_prob_rew'].sum()
merged_df['sim_prob_rew'] = merged_df['sim_prob_rew'] / total_prob * 100
sub_ebike_trips_m1_rew = merged_df[['mode_x', 'sim_prob', 'sim_prob_rew']]
sub_ebike_trips_m1_rew.rename(columns={'mode_x': 'mode'}, inplace=True)
sub_ebike_trips_m1_rew.to_csv('../validation/substitution_ebike_trips_m1_rew.csv')
sub_ebike_trips_m1_rew

# Reweight dist substitution, m1
sub_ebike_dist_m1['mode_int'] = sub_ebike_dist_m1['mode'].map(mode_mapping)
merged_df = pd.merge(sub_ebike_dist_m1, reweight, left_on='mode_int', right_on='mode', how='left')
merged_df['sim_prob_length_rew'] = merged_df['sim_prob_length'] * merged_df['reweight']
total_prob = merged_df['sim_prob_length_rew'].sum()
merged_df['sim_prob_length_rew'] = merged_df['sim_prob_length_rew'] / total_prob * 100
sub_ebike_dist_m1_rew = merged_df[['mode_x', 'sim_prob_length', 'sim_prob_length_rew']]
sub_ebike_dist_m1_rew.rename(columns={'mode_x': 'mode'}, inplace=True)
sub_ebike_dist_m1_rew.to_csv('../validation/substitution_ebike_dist_m1_rew.csv')
sub_ebike_dist_m1_rew

# Reweight trip substitution, m2
sub_ebike_trips_m2['mode_int'] = sub_ebike_trips_m2['mode'].map(mode_mapping)
merged_df = pd.merge(sub_ebike_trips_m2, reweight, left_on='mode_int', right_on='mode', how='left')
merged_df['sim_prob_rew'] = merged_df['sim_prob'] * merged_df['reweight']
total_prob = merged_df['sim_prob_rew'].sum()
merged_df['sim_prob_rew'] = merged_df['sim_prob_rew'] / total_prob * 100
sub_ebike_trips_m2_rew = merged_df[['mode_x', 'sim_prob', 'sim_prob_rew']]
sub_ebike_trips_m2_rew.rename(columns={'mode_x': 'mode'}, inplace=True)
sub_ebike_trips_m2_rew.to_csv('../validation/substitution_ebike_trips_m2_rew.csv')
sub_ebike_trips_m2_rew

# Reweight dist substitution, m2
sub_ebike_dist_m2['mode_int'] = sub_ebike_dist_m2['mode'].map(mode_mapping)
merged_df = pd.merge(sub_ebike_dist_m2, reweight, left_on='mode_int', right_on='mode', how='left')
merged_df['sim_prob_length_rew'] = merged_df['sim_prob_length'] * merged_df['reweight']
total_prob = merged_df['sim_prob_length_rew'].sum()
merged_df['sim_prob_length_rew'] = merged_df['sim_prob_length_rew'] / total_prob * 100
sub_ebike_dist_m2_rew = merged_df[['mode_x', 'sim_prob_length', 'sim_prob_length_rew']]
sub_ebike_dist_m2_rew.rename(columns={'mode_x': 'mode'}, inplace=True)
sub_ebike_dist_m2_rew.to_csv('../validation/substitution_ebike_dist_m2_rew.csv')
sub_ebike_dist_m2_rew
